+++ code/src/cljs/sixsq/nuvla/ui/edges_detail/views.cljs
@@ -1757,10 +1757,9 @@
                            :required true}]
             "Shell script: "
             [uix/EditorShell {:value     ""
-                              :on-change (fn [_editor _data value]
+                              :on-change (fn [value]
                                            (ui-callback/input-callback
-                                             (swap! form-data assoc :run value)))
-                              :class     "full-width"}]]]]
+                                             (swap! form-data assoc :run value)))}]]]]
          [ui/ModalActions
           [uix/Button
            {:text     (@tr [:add])
@@ -1868,12 +1867,11 @@
 
                    [ui/Container {:text-align "left"}
                     "Shell script: "
-                    [uix/EditorShell {:value      (:run @selected-playbook)
-                                      :on-change  (fn [_editor _data value]
-                                                    (reset! run value)
-                                                    (reset! run-changed? true))
-                                      :read-only? (not @can-edit?)
-                                      :class "full-height"}]]
+                    [uix/EditorShell {:value     (:run @selected-playbook)
+                                      :on-change (fn [value]
+                                                   (reset! run value)
+                                                   (reset! run-changed? true))
+                                      :editable  @can-edit?}]]
 
                    [uix/Button {:primary  true
                                 :text     (@tr [:save])